调整数据结构

+++ src/main/java/cn/ken/shoes/model/entity/PoisonPriceDO.java
@@ -18,9 +18,9 @@ public class PoisonPriceDO {
 
     private String euSize;
 
-    private BigDecimal normalPrice;
+    private Integer normalPrice;
 
-    private BigDecimal lightningPrice;
+    private Integer lightningPrice;
 
 
 }
